Add comment
All benches are ready, the only thing left to do is to run them properly on the server.
Energy generation benches remain unfinished.
I thought it might be a good idea to rewrite them using v2 benchmarking module, so that
the code analyzer would be able to tell what are the problems. But this approach can
be ommited